backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg[binary]==3.2.10
python-telegram-bot==20.7
python-dotenv==1.0.0
